+++ lambdas/api/index.js
@@ -1,9 +1,10 @@
 const satlib = require('../../libs')
 const logger = console
+const httpMethods = require('../../utils/http-methods')
 
 
 function determineEndpoint(event) {
-  let endpoint = process.env.SATAPI_URL
+  let endpoint = process.env.STAC_ROOT_URL
   if (typeof endpoint === 'undefined') {
     if ('X-Forwarded-Host' in event.headers) {
       endpoint = `${event.headers['X-Forwarded-Proto']}://${event.headers['X-Forwarded-Host']}`
@@ -20,10 +21,10 @@ function determineEndpoint(event) {
 function buildRequest(event) {
   const method = event.httpMethod
   let query = {}
-  if (method === 'POST' && event.body) {
-    query = JSON.parse(event.body)
-  } else if (method === 'GET' && event.queryStringParameters) {
+  if (method === httpMethods.GET && event.queryStringParameters) {
     query = event.queryStringParameters
+  } else if (event.body) {
+    query = JSON.parse(event.body)
   }
   return query
 }
@@ -44,7 +45,7 @@ module.exports.handler = async (event) => {
   logger.debug(`Event: ${JSON.stringify(event)}`)
   const endpoint = determineEndpoint(event)
   const query = buildRequest(event)
-  const result = await satlib.api.API(event.path, query, satlib.es, endpoint)
+  const result = await satlib.api.API(event.path, query, satlib.es, endpoint, event.httpMethod)
   return result instanceof Error ?
     buildResponse(404, result.message) :
     buildResponse(200, JSON.stringify(result))