+++ lambdas/ingest/index.js
@@ -1,13 +1,15 @@
 'use strict'
-
-const satlib = require('../../libs')
+const AWS = require('aws-sdk')
+const esClient = require('../../libs/esClient.js')
+const stream = require('../../libs/esStream.js')
+const ingest = require('../../libs/ingest.js')
 const logger = console
-
 
 module.exports.handler = async function handler(event) {
   logger.debug(`Event: ${JSON.stringify(event)}`)
   if (event.create_indices) {
-    await satlib.es.create_indices()
+    await esClient.create_index('collections')
+    //await satlib.es.create_index('items')
   }
 
   // start with message as is
@@ -23,14 +25,13 @@ module.exports.handler = async function handler(event) {
       logger.debug(`Record: ${JSON.stringify(i)}`)
       return i
     }))
-  }
 
-  try {
-    logger.info(`Ingesting ${items.length} items`)
-    await satlib.ingest.ingestItems(items, satlib.es)
-    logger.debug(`Ingested ${items.length} Items: ${JSON.stringify(items)}`)
-  } catch (error) {
-    console.log(error)
-    throw(error)
+    try {
+      await ingest.ingestItems(items, stream)
+      logger.info(`Ingested ${items.length} Items: ${JSON.stringify(items)}`)
+    } catch (error) {
+      console.log(error)
+      throw (error)
+    }
   }
 }