+++ libs/index.js
@@ -5,5 +5,6 @@
 module.exports = {
   api: require('./api.js'),
   es: require('./es.js'),
+  esStream: require('./esStream.js'),
   ingest: require('./ingest.js')
 }